Moving test filters around
