+++ test/integration/BUILD
@@ -190,33 +190,6 @@ envoy_cc_test_library(
 )
 
 envoy_cc_test_library(
-    name = "pass_through_filter_lib",
-    srcs = [
-        "pass_through_filter.h",
-    ],
-    deps = [
-        "//include/envoy/http:filter_interface",
-        "//include/envoy/registry",
-    ],
-)
-
-envoy_cc_test_library(
-    name = "add_trailers_filter_config_lib",
-    srcs = [
-        "add_trailers_filter.cc",
-        "add_trailers_filter.h",
-        "add_trailers_filter_config.cc",
-        "add_trailers_filter_config.h",
-    ],
-    deps = [
-        "//include/envoy/http:filter_interface",
-        "//include/envoy/registry",
-        "//include/envoy/server:filter_config_interface",
-        "//source/extensions/filters/http/common:empty_http_filter_config_lib",
-    ],
-)
-
-envoy_cc_test_library(
     name = "http_integration_lib",
     srcs = [
         "http_integration.cc",
@@ -225,15 +198,15 @@ envoy_cc_test_library(
         "http_integration.h",
     ],
     deps = [
-        ":add_trailers_filter_config_lib",
         ":integration_lib",
-        ":pass_through_filter_lib",
         ":test_host_predicate_lib",
         "//include/envoy/event:timer_interface",
         "//source/common/common:thread_annotations",
         "//source/extensions/filters/http/router:config",
         "//source/extensions/filters/network/http_connection_manager:config",
         "//test/common/upstream:utility_lib",
+        "//test/integration/filters:add_trailers_filter_config_lib",
+        "//test/integration/filters:pause_filter_lib",
         "//test/test_common:registry_lib",
     ],
 )
