--- test/integration/add_trailers_filter.h
@@ -1,44 +1,0 @@
-#pragma once
-
-#include "envoy/http/filter.h"
-
-namespace Envoy {
-// a test filter that inserts trailers at the end of encode/decode
-class AddTrailersStreamFilter : public Http::StreamFilter {
-public:
-  // Http::StreamFilterBase
-  void onDestroy() override {}
-
-  // Http::StreamDecoderFilter
-  Http::FilterHeadersStatus decodeHeaders(Http::HeaderMap&, bool) override {
-    return Http::FilterHeadersStatus::Continue;
-  }
-  Http::FilterDataStatus decodeData(Buffer::Instance&, bool end_stream) override;
-
-  Http::FilterTrailersStatus decodeTrailers(Http::HeaderMap&) override {
-    return Http::FilterTrailersStatus::Continue;
-  }
-  void setDecoderFilterCallbacks(Http::StreamDecoderFilterCallbacks& callbacks) override {
-    decoder_callbacks_ = &callbacks;
-  }
-
-  // Http::StreamEncoderFilter
-  Http::FilterHeadersStatus encode100ContinueHeaders(Http::HeaderMap&) override {
-    return Http::FilterHeadersStatus::Continue;
-  }
-  Http::FilterHeadersStatus encodeHeaders(Http::HeaderMap&, bool) override {
-    return Http::FilterHeadersStatus::Continue;
-  }
-  Http::FilterDataStatus encodeData(Buffer::Instance&, bool end_stream) override;
-  Http::FilterTrailersStatus encodeTrailers(Http::HeaderMap&) override {
-    return Http::FilterTrailersStatus::Continue;
-  }
-  void setEncoderFilterCallbacks(Http::StreamEncoderFilterCallbacks& callbacks) override {
-    encoder_callbacks_ = &callbacks;
-  }
-
-private:
-  Http::StreamDecoderFilterCallbacks* decoder_callbacks_{};
-  Http::StreamEncoderFilterCallbacks* encoder_callbacks_{};
-};
-} // namespace Envoy
